requirements.txt
hatchling
pydantic-ai-slim[openai,vertexai,groq,anthropic]==0.0.20
asyncpg>=0.30.0
fastapi>=0.115.4
logfire[asyncpg,fastapi,sqlite3]>=2.6
python-multipart>=0.0.17
rich>=13.9.2
uvicorn>=0.32.0
devtools>=0.12.2
gradio>=5.9.0; python_version>'3.9'